test(release): cover dry-run cleanup
